Add Cassandra connection checking to ping.
When it gets a bad connection, telephus can sit on Cassandra requests forever
and never time them out. This commit adds logic to the ping handler to
attempt to send a request on each Cassandra connection. If this fails,
the ping will time out, and the crest process will be restarted by
monit.